fsck: move exfat_de_iter functions to de_iter.c
