+++ fsck/Makefile.am
@@ -3,4 +3,4 @@ fsck_exfat_LDADD = $(top_builddir)/lib/libexfat.a
 
 sbin_PROGRAMS = fsck.exfat
 
-fsck_exfat_SOURCES = fsck.c repair.c fsck.h repair.h
+fsck_exfat_SOURCES = fsck.c repair.c fsck.h de_iter.c repair.h
